Create a new Dotto

It creates a Dotto (.Rmd) template based on a title and engine language.

create_Dotto(title = NULL, lang = "r", open = T)

Arguments

title
Title of the dotto post.

lang
The programming language for writing the Dotto. Default is R. Other options are python, julia, sql, bash, js, node, d3, Rcpp, stan. It is not case-sensitive.

open
If open the dotto file in Rstudio or not. Default is TRUE.

Details

For the yml metadata, here are more details about some tags:

lang: R, python, julia, sql, bash, node, etc.

categories: Importing", "Cleaning, Manipulation, Visualization, Machine-Learning, Statistical-Learning, Reporting, Data-Engineering

site: A personal url address, e.g., a personal website, twitter account, linkedin, etc.

img_url: A canonical link into your img that should be saved under _files dir.

applications: Sector specific application require domain knowledge expertise like Finance, Healthcare, Insurance, etc.

cover_image: An external/internal link to the img that will be displayed in the cards.