Update Dotto file and folder names based on updated slug

The Dotto url is created based on the name on the Dotto folder, which follows the format D#-date-slug, e.g. D112-2020-11-21-data-sampling. As the Dotto file and folder names are created originally by slug, for changing the file/folder names, it is recommended to change slug first, and then run update_Dotto().

update_Dotto(path = NULL, clean_dotto_deps = T)

Arguments

path
a path to the Dotto .Rmd file. IF the path is a rmd folder path, it automatically find the Rmd path if there is only one unique Rmd file in the folder.

clean_Dotto_deps
Delete all the files and folders in Dotto and only keep .Rmd file. Default is FALSE.

Details

for creating a valid slug, you can use DataMotto::create_slug("foo")